Feature/basket controller

+++ PraisePoint/Services/Basket/Basket.API/Controllers/BasketController.cs
@@ -1,5 +1,11 @@
-﻿using Basket.API.Entities;
+﻿using System.Security.Claims;
+using AutoMapper;
+using Basket.API.Entities;
+//using Basket.API.GrpcServices;
 using Basket.API.Repositories;
+using EventBus.Messages.Events;
+using Grpc.Core;
+using MassTransit;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Basket.API.Controllers
@@ -9,16 +15,32 @@ namespace Basket.API.Controllers
     public class BasketController : ControllerBase
     {
         private readonly IBasketRepository _repository;
+        //private readonly CouponGrpcService _couponGrpcService;
+        private readonly ILogger<BasketController> _logger;
+        private readonly IMapper _mapper;
+        private readonly IPublishEndpoint _publishEndpoint;
 
-        public BasketController(IBasketRepository repository)
+        public BasketController(IBasketRepository repository,//CouponGrpcService couponGrpcService,
+        ILogger<BasketController> logger,
+        IMapper mapper,
+        IPublishEndpoint publishEndpoint)
         {
             _repository = repository ?? throw new ArgumentNullException(nameof(repository));
+            // _couponGrpcService = couponGrpcService ?? throw new ArgumentNullException(nameof(couponGrpcService));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _mapper = mapper ?? throw new ArgumentNullException(nameof(logger));
+            _publishEndpoint = publishEndpoint ?? throw new ArgumentNullException(nameof(logger));
         }
 
         [HttpGet("{username}")]
         [ProducesResponseType(typeof(ShoppingCart), StatusCodes.Status200OK)]
         public async Task<ActionResult<ShoppingCart>> GetBasket(string username)
         {
+            if (User.FindFirst(ClaimTypes.Name)?.Value != username)
+            {
+                return Forbid();
+            }
+
             var basket = await _repository.GetBasket(username);
             return Ok(basket ?? new ShoppingCart(username));
         }
@@ -27,15 +49,50 @@ namespace Basket.API.Controllers
         [ProducesResponseType(typeof(ShoppingCart), StatusCodes.Status200OK)]
         public async Task<ActionResult<ShoppingCart>> UpdateBasket([FromBody] ShoppingCart basket)
         {
+            if (User.FindFirstValue(ClaimTypes.Name) != basket.Username)
+            {
+                return Forbid();
+            }
+
             return Ok(await _repository.UpdateBasket(basket));
         }
 
         [HttpDelete("{username}")]
         [ProducesResponseType(typeof(void), StatusCodes.Status200OK)]
-        public async Task<ActionResult> DeleteBasket(string username)
+        public async Task<IActionResult> DeleteBasket(string username)
         {
+            if (User.FindFirstValue(ClaimTypes.Name) != username)
+            {
+                return Forbid();
+            }
+
             await _repository.DeleteBasket(username);
             return Ok();
+        }
+
+        [Route("[action]")]
+        [HttpPost]
+        [ProducesResponseType(typeof(void), StatusCodes.Status202Accepted)]
+        [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> Checkout([FromBody] BasketCheckout basketCheckout)
+        {
+            if (User.FindFirstValue(ClaimTypes.Name) != basketCheckout.BuyerUsername)
+            {
+                return Forbid();
+            }
+
+            var basket = await _repository.GetBasket(basketCheckout.BuyerUsername);
+            if (basket is null)
+            {
+                return BadRequest();
+            }
+
+            var eventMessage = _mapper.Map<BasketCheckoutEvent>(basketCheckout);
+            await _publishEndpoint.Publish(eventMessage);
+
+            await _repository.DeleteBasket(basketCheckout.BuyerUsername);
+
+            return Accepted();
         }
     }
 }
